修改强转uid类型

+++ frameworks/js/napi/netpolicy/src/netpolicy_exec.cpp
@@ -169,7 +169,7 @@ bool NetPolicyExec::ExecGetDeviceIdleAllowList(GetDeviceIdleAllowListContext *co
 bool NetPolicyExec::ExecSetPowerSaveAllowList(SetPowerSaveAllowListContext *context)
 {
     int32_t result =
-        DelayedSingleton<NetPolicyClient>::GetInstance()->SetPowerSaveAllowedList(context->uid_, context->isAllow_);
+        DelayedSingleton<NetPolicyClient>::GetInstance()->SetPowerSaveAllowedList(std::static_cast<uint32_t>(context->uid_), context->isAllow_);
     if (result != NETMANAGER_SUCCESS) {
         NETMANAGER_BASE_LOGE("ExecSetPowerSaveAllowList error: uid = %{public}d, result = %{public}d",
                              context->uid_, result);
